requirements.txt
pytube
python-telegram-bot